refactor: remove console logs and improve user script management
- Removed unnecessary console log statements across various components for cleaner code.
- Introduced built-in user scripts for YouTube ad skipping and SponsorBlock functionality.
- Enhanced user script management by allowing enabling/disabling of built-in scripts.
- Updated user script model to include a builtIn property for better identification.
- Improved the persistence and loading logic for user scripts to maintain state.
