+++ appData/interface.json
@@ -1,1 +1,1 @@
-{"layout":"BASIC","mode":"light","dataSync":{"intervalTime":"30","hardwareAcceleration":"1"},"extension":{"adblock":true,"vault":true,"translate":true,"userscript":true,"disabledFilters":["21442ad94c9864d18fbfe288a5dbc559","f4c9ccf98ce5d9d6e432058ad93b7ad7"]},"savedCookies":"0","historyRetentionDays":"30","hibernateMode":"normal","hibernateCustomMinutes":60}
+{"layout":"BASIC","mode":"light","dataSync":{"intervalTime":"30","hardwareAcceleration":"1"},"extension":{"adblock":false,"vault":true,"translate":true,"userscript":true,"disabledFilters":["21442ad94c9864d18fbfe288a5dbc559","f4c9ccf98ce5d9d6e432058ad93b7ad7"]},"savedCookies":"0","historyRetentionDays":"30","hibernateMode":"normal","hibernateCustomMinutes":60}
