+++ src/features/adblocker/index.ts
@@ -1,2 +1,3 @@
 export * as AdblockController from "./controllers";
 export * as AdblockService from "./services";
+export * from "./scripts";
