+++ src/features/ui/components/tabGroup/TabGroupHeader.tsx
@@ -33,7 +33,7 @@ const TabGroupHeader = memo(({ group, tabCount, onContextMenu }: TabGroupHeaderP
 
   return (
     <div
-      className="flex items-center gap-1 px-1 rounded-md shrink-0 flex-col"
+      className="flex items-center gap-1 px-1 rounded-md shrink-0 flex-col z-2"
       onContextMenu={(e) => {
         e.preventDefault();
         onContextMenu?.(e, group.id);
@@ -49,36 +49,36 @@ const TabGroupHeader = memo(({ group, tabCount, onContextMenu }: TabGroupHeaderP
           <button
             type="button"
             onClick={editGroup}
-            className="w-3.5 h-3.5 flex items-center justify-center text-slate-400 hover:text-indigo-500 rounded shrink-0"
+            className="flex items-center justify-center text-slate-400 hover:text-slate-800 rounded shrink-0 cursor-pointer"
           >
-            <IconPencil size={10} />
+            <IconPencil size={14} />
           </button>
           <button
             type="button"
             onClick={(e) => { e.stopPropagation(); deleteGroup(); }}
-            className="w-3.5 h-3.5 flex items-center justify-center text-slate-400 hover:text-red-500 rounded shrink-0"
+            className="flex items-center justify-center text-slate-400 hover:text-red-500 rounded shrink-0 cursor-pointer"
           >
-            <IconTrash size={10} />
+            <IconTrash size={14} />
           </button>
         </div>
       </div>
 
       <div className="flex items-center gap-0.5 shrink-0">
-        {group.hidden && <IconEyeOff size={10} className="text-slate-300" />}
+        {group.hidden && <IconEyeOff size={16} className="text-slate-300" />}
         <button
           type="button"
           onClick={(e) => { e.stopPropagation(); toggleHide(); }}
           className="w-3.5 h-3.5 flex items-center justify-center text-indigo-500 hover:text-indigo-800 rounded shrink-0 cursor-pointer"
           title={group.hidden ? "Show group" : "Hide group from sidebar"}
         >
-          {group.hidden ? <IconEye size={10} /> : <IconEyeOff size={10} />}
+          {group.hidden ? <IconEye size={16} /> : <IconEyeOff size={16} />}
         </button>
-        <span
+        {/* <span
           className="inline-flex items-center justify-center min-w-[18px] h-[16px] px-1 rounded-full text-[9px] font-bold"
           style={{ backgroundColor: `${group.color}20`, color: group.color }}
         >
           {tabCount}
-        </span>
+        </span> */}
       </div>
     </div>
   );
