+++ src/features/userscript/controllers/index.ts
@@ -8,6 +8,30 @@ import { cacheSystem } from "~/features/cacheSystem";
 import { UserScriptService } from "../services";
 import { eventStore } from "~/core/stores";
 import { WebContentsView } from "electron";
+import { SkipADSBlock, SponsorBlock } from "~/features/adblocker/scripts";
+
+const BUILT_IN_SCRIPTS: IUserScript[] = [
+  {
+    id: "builtin-skip-adsblock",
+    name: "YouTube Ad Skip",
+    source: `if (!window.__ytAdblockInjected) { (${SkipADSBlock.toString()})(); }`,
+    matches: ["*://*.youtube.com/*"],
+    excludes: [],
+    runAt: "document-end",
+    enabled: true,
+    builtIn: true,
+  },
+  {
+    id: "builtin-sponsorblock",
+    name: "SponsorBlock",
+    source: `if (!window.__ytAdblockInjected) { (${SponsorBlock.toString()})(); }`,
+    matches: ["*://*.youtube.com/*"],
+    excludes: [],
+    runAt: "document-end",
+    enabled: true,
+    builtIn: true,
+  },
+];
 
 export class UserScriptController {
   private store: StoreManager = new StoreManager("userscripts");
@@ -15,10 +39,24 @@ export class UserScriptController {
   private initialized = false;
   private initializing?: Promise<void>;
   activeTab: WebContentsView | null = null;
+  private disabledBuiltIns: Set<string> = new Set();
   constructor(private service: UserScriptService = new UserScriptService()) {
     eventStore.listen("viewChanges", (view: WebContentsView) => {
       this.activeTab = view;
     });
+  }
+
+  disableBuiltIn(id: string) {
+    this.disabledBuiltIns.add(id);
+  }
+
+  enableBuiltIn(id: string) {
+    this.disabledBuiltIns.delete(id);
+  }
+
+  setBuiltInEnabled(id: string, enabled: boolean) {
+    if (enabled) this.enableBuiltIn(id);
+    else this.disableBuiltIn(id);
   }
   async initialize() {
     if (this.initialized) return;
@@ -30,15 +68,32 @@ export class UserScriptController {
   private async load() {
     const callback = () => this.store.readFiles<IUserScriptStore>();
     const raw = await cacheSystem.get<IUserScriptStore>("userscripts", callback);
-    const scripts = Array.isArray(raw?.scripts) ? raw.scripts : [];
-    this.scripts = new Map(scripts.filter((script) => script?.id).map((script) => [script.id, new UserScript(script)]));
+    const storedScripts = Array.isArray(raw?.scripts) ? raw.scripts : [];
+    const storedMap = new Map(storedScripts.filter((s) => s?.id).map((s) => [s.id, s]));
+
+    // Load built-in scripts from code, preserving toggle state from storage
+    this.scripts = new Map();
+    for (const builtIn of BUILT_IN_SCRIPTS) {
+      const stored = storedMap.get(builtIn.id);
+      this.scripts.set(
+        builtIn.id,
+        new UserScript({ ...builtIn, enabled: stored ? stored.enabled : builtIn.enabled }),
+      );
+    }
+
+    // Load user scripts from storage
+    for (const [id, s] of storedMap) {
+      if (BUILT_IN_SCRIPTS.some((b) => b.id === id)) continue;
+      this.scripts.set(id, new UserScript(s));
+    }
+
     this.initialized = true;
   }
 
   private persist() {
-    const scripts = this.listScripts();
-    cacheSystem.set("userscripts", { scripts });
-    return this.store.saveFiles({ scripts });
+    const allScripts = [...this.scripts.values()].map((script) => script.toJSON());
+    cacheSystem.set("userscripts", { scripts: allScripts });
+    return this.store.saveFiles({ scripts: allScripts });
   }
 
   listScripts() {
@@ -52,6 +107,16 @@ export class UserScriptController {
     }
     const now = Date.now();
     const current = id ? this.scripts.get(id) : undefined;
+
+    // For built-in scripts, only toggle the enabled state
+    if (current?.builtIn) {
+      current.enabled = typeof data.enabled === "boolean" ? data.enabled : current.enabled;
+      current.updatedAt = Date.now();
+      this.scripts.set(current.id, current);
+      this.persist();
+      return current;
+    }
+
     const script: UserScript = new UserScript({
       id: current?.id || uuid_v7(),
       ...data,
@@ -80,6 +145,10 @@ export class UserScriptController {
 
   async deleteScript(id: string) {
     await this.initialize();
+    const script = this.scripts.get(id);
+    if (script?.builtIn) {
+      throw new Error("Cannot delete built-in scripts");
+    }
     this.scripts.delete(id);
     this.persist();
     return true;
@@ -100,6 +169,7 @@ export class UserScriptController {
     await this.initialize();
     return this.listScripts().filter((script) => {
       if (!script.enabled) return false;
+      if (script.builtIn && this.disabledBuiltIns.has(script.id)) return false;
       const isMatched = isUrlMatchedByPatterns(url, script.matches);
       if (!isMatched) return false;
       const isExcluded = script.excludes?.length ? isUrlMatchedByPatterns(url, script.excludes) : false;
@@ -114,14 +184,20 @@ export class UserScriptController {
     let newScripts = Array.isArray(results) ? results : [];
     if (newScripts.length) {
       newScripts = newScripts.filter((script) => script?.id);
-      // this.scripts = new Map(
-      //   newScripts.filter((script) => script?.id).map((script) => [script.id, new UserScript(script)]),
-      // );
-      // this.scripts = new Map(newScripts.map((script) => [script.id, new UserScript(script)]));
-      this.scripts = new Map();
-      for (let script of newScripts) {
-        await this.saveScript(script);
+      // Keep built-in scripts, update only their enabled state from manager results
+      const builtIn = [...this.scripts.entries()].filter(([, s]) => s.builtIn);
+      this.scripts = new Map(builtIn);
+      for (const s of newScripts) {
+        if (s.builtIn) {
+          const existing = this.scripts.get(s.id);
+          if (existing) {
+            existing.enabled = typeof s.enabled === "boolean" ? s.enabled : existing.enabled;
+          }
+        } else {
+          this.scripts.set(s.id, new UserScript(s));
+        }
       }
+      this.persist();
     }
   }
 }
