+++ src/features/userscript/plugin/index.ts
@@ -42,7 +42,6 @@ export class UserScriptTabPlugin implements ITabPlugin {
             })();
           `;
           ctx.webContents.executeJavaScript(codeToInject, true).catch((err) => console.error("Execution failed:", err));
-          // console.log(`[UserScript:${script.name}] executed (runAt=${runAt}) on ${url}`);
         } catch (error) {
           console.error(`[UserScript:${script.name}] execution failed`, error);
         }
