+++ src/features/userscript/services/index.ts
@@ -60,8 +60,6 @@ export class UserScriptService {
       return new Promise<any[] | null>((resolve, reject) => {
         let isReady = false;
         const onConsoleMessage = (_event: any, _level: any, message: string) => {
-          // console.log("onConsoleMessage level", _level);
-          // console.log("onConsoleMessage message", message);
           if (!message.startsWith(SENTINEL)) return;
           try {
             const json = message.slice(SENTINEL.length);
