+++ src/features/userscript/types/userscript.d.ts
@@ -10,6 +10,7 @@ export interface IUserScript {
   excludes?: string[];
   runAt: UserScriptRunAt;
   enabled: boolean;
+  builtIn?: boolean;
   createdAt?: number;
   updatedAt?: number;
 }
